Reordering tasks to kill everything on each run
Experienced an issue where mocha throwing errors would cause things to not shut down. Also, there's a weird issue with invoking http from within a gulp task that can hold the response stream open and cause gulp to hang without exiting... so the selenium task has to force the stream on the HTTP query to end before hitting the callback.